+++ src/Libraries/SmartStore.Core/Domain/Catalog/CatalogSettings.cs
@@ -38,6 +38,8 @@ namespace SmartStore.Core.Domain.Catalog
 			ProductSearchAutoCompleteNumberOfProducts = 10;
 			ProductSearchTermMinimumLength = 3;
 			NumberOfBestsellersOnHomepage = 6;
+            ShowManufacturersOnHomepage = true;
+            ShowManufacturerPictures = false;
 			SearchPageProductsPerPage = 6;
 			ProductsAlsoPurchasedEnabled = true;
 			ProductsAlsoPurchasedNumber = 6;
@@ -282,6 +284,16 @@ namespace SmartStore.Core.Domain.Catalog
         /// Gets or sets a number of bestsellers on home page
         /// </summary>
         public int NumberOfBestsellersOnHomepage { get; set; }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether to show manufacturers on home page
+        /// </summary>
+        public bool ShowManufacturersOnHomepage { get; set; }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether to show manufacturer pictures or names on home page
+        /// </summary>
+        public bool ShowManufacturerPictures { get; set; }
 
         /// <summary>
         /// Gets or sets a number of products per page on search products page