+++ src/Libraries/SmartStore.Data/Migrations/201506181858349_AclRecordCustomerRole.cs
@@ -119,6 +119,18 @@ namespace SmartStore.Data.Migrations
 			builder.AddOrUpdate("Admin.Configuration.Payment.Methods.Fields.RecurringPaymentType",
 				"Recurring payments",
 				"Wiederkehrende Zahlungen");
+
+            builder.AddOrUpdate("Admin.Configuration.Settings.Catalog.ShowManufacturersOnHomepage",
+                "Show manufacturers on homepage",
+                "Zeige Hersteller auf der Homepage",
+                "Specifies whether manufacturers will be displayed on the homepage.",
+                "Bestimmt ob Hersteller auf der Homepage angezeigt werden.");
+
+            builder.AddOrUpdate("Admin.Configuration.Settings.Catalog.ShowManufacturerPictures",
+                "Show manufacturer pictures on homepage",
+                "Zeige Herstellerbilder auf der Homepage",
+                "Specifies whether manufacturers will be displayed as images or textual links on the homepage.",
+                "Bestimmt ob Hersteller auf der Homepage als Bilder oder textuelle Links angezeigt werden.");
         }
     }
 }